+++ package.json
@@ -15,6 +15,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-icons": "^5.5.0",
+    "react-router-dom": "^7.4.0",
     "zod": "^3.24.2"
   },
   "devDependencies": {
